lets make node ids bigger:

+++ src/node.h
@@ -8,16 +8,15 @@
 
 struct _cassie_body_id_t_
 {
-    uint8_t id;
+    uint16_t id;
 };
 typedef struct _cassie_body_id_t_ cassie_body_id_t;
 
 struct _node_body_id_t_
 {
-    uint8_t id;
+    uint16_t id;
 };
 typedef struct _node_body_id_t_ node_body_id_t;
-
 
 
 typedef mjtNum* v3_t;
